requirements.txt
tensorflow>=2.13.0
gradio>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
